+++ src/utils/index.ts
@@ -3,5 +3,6 @@
  */
 
 export * from './dom';
+export * from './text-utils';
 
 export { log, createLogger } from '../core/logger';